week9/json_sorter/requirements/requirements.prod.txt
pydantic==1.9.0
python-decouple==3.6
gunicorn==20.1.0
pymongo==4.1.0
orjson==3.6.7
flask==2.1.1
fnvhash==0.1.0